dll文件不存在重复处理问题

+++ Just.ExtractTfs/ExtractTfsVM.cs
@@ -539,7 +539,7 @@ namespace Just.ExtractTfs
                         }
                     }
                 }
-                else
+                else if (!File.Exists((path + "NotExists")))
                 {
                     File.Create(path + "NotExists").Close();
                 }
